Outline of Physics has been created
The idea is to have a thread for physics and input so that they aren't
linked to rendering. This way, if there is a big simulation that needs
to happen, like an expolsion, the rendering loop isn't effected by it.
The main problem will be syncing them up to avoid concurrent
modification problems. Experimentation begins now.