Improved GUI Element structure
Added a hashmap to the GUI class so that all elements are easily
searchable and managable, makes refrencing them a lot cleaner. Also
updated ClickableElement class to change the color of the text inside
the class instead of having to write methods to do so every time you use
it.