Upgrades appear to be working smoothly
Added a test upgrade to make the player jump higher when the upgrade is
in use
Upgrade has a life span and a rate at which it runs out
Upgrade can hold either one value or multiple to have more effect on the
player
Upgrades are managed like Entities, when their life runs out, they are
removed from the list. They are easily added to a large batch and then
effect the Entities as such. Works pretty smooth, but some of the
implementation is convoluted, will need to polish that up at some point.