Made the upgrade system work smoother
Now instead of creating a new GUI for every upgrade, it just resubmits
the Upgrade data to the overlay, which speeds up the whole process.
Before, it dropped frames when jumping on the button because it had to
load a new font since a new GUI was being created, but now the GUI is
pre-created so that it doesn't need to load everything every time you
hit the button. Works much better now.