New Effects
Added a new effects system.
Works off the same principal as the entity system, but I needed
different features for the Effects system.
This was used to add the stars in the background
The render call is split into two parts, front and back. This allows
things to be rendered either in front or behind objects depending on
it's Z location set in the initializer method