New Upgrade Value System Implemented
Now all of an entities move characteristics are stored in a HashMap, and
upgrades then add an extra amount onto a specific value from that list.
They are now also capable of effecting more than one value, as many as
they want really. And instead of setting a new value, they sum together,
so the upgrade is basically an offset from the current number.